backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
